Refactored settings page to use daisyui and moved manage characters table into separate component
