+++ src/app/pages/settings/settings.module.ts
@@ -9,9 +9,21 @@ import { AddCharacterFormComponent } from './components/add-character-form/add-c
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
 import { TippyModule } from '@ngneat/helipopper';
 import { NgSelectModule } from '@ng-select/ng-select';
+import { LayoutModule } from 'src/app/layout/layout.module';
+import { ManageCharactersComponent } from './components/manage-characters/manage-characters.component';
+import { AboutComponent } from './components/about/about.component';
 
 @NgModule({
-  declarations: [SettingsComponent, AddCharacterFormComponent],
-  imports: [CommonModule, SettingsRoutingModule, ReactiveFormsModule, NgSelectModule, FontAwesomeModule, TippyModule, FrameworkModule],
+  declarations: [SettingsComponent, AddCharacterFormComponent, ManageCharactersComponent, AboutComponent],
+  imports: [
+    CommonModule,
+    SettingsRoutingModule,
+    ReactiveFormsModule,
+    NgSelectModule,
+    FontAwesomeModule,
+    TippyModule,
+    FrameworkModule,
+    LayoutModule,
+  ],
 })
 export class SettingsModule {}
